[Refactor] Replace TokensCardCollectionCoordinator with TokensCardCoordinator #4493
